requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==5.3.2
certifi==2025.4.26
cffi==1.16.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.4
duckdb==1.4.1
Flask==3.1.1
Flask-Bcrypt==1.0.1
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
javascript==1!1.1.2
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
nextcord==2.6.0
oauthlib==3.2.2
packaging==23.2
pendulum==3.1.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.6.0
PyNaCl==1.4.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==1.3.1
rsa==4.9
setuptools==78.1.1
six==1.17.0
typing-inspect==0.9.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.5.0
waitress==3.0.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.9.4
